requirements.txt
pyusb>=1.2.1
libusb-package>=1.0.26.2
fastapi>=0.110
uvicorn[standard]>=0.27
anthropic>=0.40
